requirements.txt
asgiref==3.7.2
Django==4.2.4
django-dotenv==1.4.2
django-htmx==1.16.0
django-phonenumber-field==7.1.0
gunicorn==21.2.0
packaging==23.1
phonenumberslite==8.13.19
psycopg2==2.9.7
sqlparse==0.4.4
tzdata==2023.3
